+++ src/main/java/chessgame/domain/chessgame/Board.java
@@ -1,4 +1,4 @@
-package chessgame.domain.board;
+package chessgame.domain.chessgame;
 
 import chessgame.domain.coordinate.Coordinate;
 import chessgame.domain.piece.Camp;
