+++ src/main/java/chessgame/domain/chessgame/DirectionVector.java
@@ -1,4 +1,4 @@
-package chessgame.domain.board;
+package chessgame.domain.chessgame;
 
 import chessgame.domain.coordinate.Coordinate;
